resolvers以下を全部バラす
